Centralizing the dependency resolution to ease transition to rosdep.
