+++ superflore/exceptions.py
@@ -1,0 +1,33 @@
+# Copyright 2017 Open Source Robotics Foundation, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+class UnknownLicense(Exception):
+    def __init__(self, message):
+        self.message = message
+
+
+class UnresolvedDependency(Exception):
+    def __init__(self, message):
+        self.message = message
+
+
+class UnknownPlatform(Exception):
+    def __init__(self, message):
+        self.message = message
+
+
+class NoPkgXml(Exception):
+    def __init__(self, message):
+        self.message = message
