+++ superflore/utils.py
@@ -13,6 +13,11 @@
 # limitations under the License.
 
 
+from superflore.exceptions import UnresolvedDependency
+from superflore.exceptions import UnknownPlatform
+from superflore.exceptions import UnknownLicense
+from termcolor import colored
+import yaml
 import sys
 import re
 
@@ -106,9 +111,11 @@ def resolve_dep(pkg, os):
     """
     if os == 'oe':
         return _resolve_dep_open_embedded(pkg)
+    elif os == 'gentoo':
+        return _resolve_dep_gentoo(pkg)
     else:
         msg = "Unknown target platform '{0}'".format(os)
-        raise UnresolvedDependency(msg)
+        raise UnknownPlatform(msg)
 
 
 def _resolve_dep_open_embedded(pkg):
@@ -132,18 +139,39 @@ def _resolve_dep_open_embedded(pkg):
         return 'python-catkin-pkg-native'
     else:
         return pkg.replace('_', '-')
-        
-
-class UnknownLicense(Exception):
-    def __init__(self, message):
-        self.message = message
 
 
-class UnresolvedDependency(Exception):
-    def __init__(self, message):
-        self.message = message
-
-
-class NoPkgXml(Exception):
-    def __init__(self, message):
-        self.message = message
+def _resolve_dep_gentoo(pkg):
+    global base_yml
+    global python_yml
+    global ruby_yml
+    if pkg not in base_yml:
+        if pkg not in python_yml:
+            if pkg not in ruby_yml:
+                raise UnresolvedDependency(
+                    "could not resolve package {} for Gentoo.".format(pkg))
+            elif 'gentoo'not in ruby_yml[pkg]:
+                raise UnresolvedDependency(
+                    "could not resolve package {} for Gentoo.".format(pkg))
+            elif 'portage' in ruby_yml[pkg]['gentoo']:
+                return ruby_yml[pkg]['gentoo']['portage']['packages'][0]
+            else:
+                resolution = ruby_yml[pkg]['gentoo'][0]
+                return resolution
+        elif 'gentoo'not in python_yml[pkg]:
+            raise UnresolvedDependency(
+                "could not resolve package {} for Gentoo.".format(pkg))
+        elif 'portage' in python_yml[pkg]['gentoo']:
+            return python_yml[pkg]['gentoo']['portage']['packages'][0]
+        else:
+            resolution = python_yml[pkg]['gentoo'][0]
+            return resolution
+    elif 'gentoo'not in base_yml[pkg]:
+        raise UnresolvedDependency(
+            "could not resolve package {} for Gentoo.".format(pkg))
+    elif 'portage' in base_yml[pkg]['gentoo']:
+        resolution = base_yml[pkg]['gentoo']['portage']['packages'][0]
+        return resolution
+    else:
+        resolution = base_yml[pkg]['gentoo'][0]
+        return resolution
